+++ test/integration/quote-all/invalid.test.js
@@ -1,5 +1,5 @@
 /**
- * @overview Contains integration tests for invalid use of `Shescape#quote`.
+ * @overview Contains integration tests for invalid use of `Shescape#quoteAll`.
  * @license MIT
  */
 
@@ -24,15 +24,36 @@ testProp(
   },
 );
 
-testProp("invalid arguments", [arbitrary.shescapeOptions()], (t, options) => {
-  let shescape;
-  try {
-    shescape = new Shescape(options);
-  } catch {
-    return t.pass();
-  }
+testProp(
+  "invalid argument list",
+  [arbitrary.shescapeOptions()],
+  (t, options) => {
+    let shescape;
+    try {
+      shescape = new Shescape(options);
+    } catch {
+      return t.pass();
+    }
 
-  for (const { value } of constants.illegalArguments) {
-    t.throws(() => shescape.quoteAll([value]), { instanceOf: TypeError });
-  }
-});
+    for (const { value } of constants.illegalArgumentLists) {
+      t.throws(() => shescape.quoteAll(value), { instanceOf: TypeError });
+    }
+  },
+);
+
+testProp(
+  "invalid individual argument",
+  [arbitrary.shescapeOptions()],
+  (t, options) => {
+    let shescape;
+    try {
+      shescape = new Shescape(options);
+    } catch {
+      return t.pass();
+    }
+
+    for (const { value } of constants.illegalArguments) {
+      t.throws(() => shescape.quoteAll([value]), { instanceOf: TypeError });
+    }
+  },
+);